Stealing opencensus standard example
